Crate tinycdb_sys

Structs

cdb

cdb_find

cdb_make

Enums

CdbFindMode

CdbPutMode
CdbPutMode represents the different behaviours that will be used when inserting a key into a database where the key already exists.

Functions

cdb_find⚠

cdb_findinit⚠

cdb_findnext⚠

cdb_free⚠

cdb_get⚠

cdb_init⚠

cdb_make_add⚠

cdb_make_exists⚠

cdb_make_find⚠

cdb_make_finish⚠

cdb_make_put⚠

cdb_make_start⚠

cdb_read⚠

cdb_seqinit⚠

cdb_seqnext⚠